refactor: extract nail comparison panel into dedicated component
- Create src/components/NailComparisonPanel.tsx
- Move comparison items sorting, rendering, and panel JSX from App.tsx
- Update src/App.tsx to integrate the new component
- Clean up unused helpers and state logic in App.tsx
